feat: swap secure credential resource to GraphQL API
